Change detector level limits to dB

+++ include/r_private.h
@@ -17,7 +17,7 @@
 #include "compat_time.h"
 
 struct dm_state {
-    int32_t level_limit;
+    float level_limit;
     int16_t am_buf[MAXIMAL_BUF_LENGTH];  // AM demodulated signal (for OOK decoding)
     union {
         // These buffers aren't used at the same time, so let's use a union to save some memory
